Changes to the design as per discussion with Jason
