+++ src/main/java/org/openpnp/spi/base/AbstractFeeder.java
@@ -32,15 +32,6 @@ public abstract class AbstractFeeder extends AbstractModelObject implements Feed
     {
         this.id = Configuration.createId();
         this.name = getClass().getSimpleName();
-        this.child = null;
-        Configuration.get().addListener(new ConfigurationListener.Adapter() {
-            @Override
-            public void configurationLoaded(Configuration configuration) throws Exception {
-                part = configuration.getPart(partId);
-                child = configuration.getMachine().getFeeder(childID);
-                parent = configuration.getMachine().getFeeder(parentID);
-            }
-        });
     }
 
     @Override
@@ -93,16 +84,5 @@ public abstract class AbstractFeeder extends AbstractModelObject implements Feed
     }
 
 
-    @Attribute(required=false)
-    protected String childID;
-    @Attribute(required=false)
-    protected String parentID;
 
-    protected Feeder child;
-    protected Feeder parent;
-    public Boolean supportsChildren() { return false; }
-    public Feeder getChild() { return null; }
-    public void setChild(Feeder child) { return; }
-    public Boolean isChildFeeder() { return false; }
-    public void setParent(Feeder parent) { return; }
 }
